backend/requirements.txt
djangorestframework
python-decouple
psycopg2-binary
pillow
drf-yasg
djangorestframework-simplejwt
django-modeltranslation
gunicorn
django-ckeditor
django-cors-headers
django-jsoneditor
